requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.0
typer[all]>=0.9.0